% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{AGE_CLAMP_MAX}
\alias{AGE_CLAMP_MAX}
\title{Age decade clamping: all ages >= this value merge into a single "80+" group.}
\usage{
AGE_CLAMP_MAX
}
\description{
Age decade clamping: all ages >= this value merge into a single "80+" group.
}
\keyword{internal}
