% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_demography.R
\name{build_pyramid}
\alias{build_pyramid}
\title{Build an interactive population pyramid}
\usage{
build_pyramid(demo_df, dataset_name)
}
\arguments{
\item{demo_df}{Demographics data frame with birth_year, sex, patient_count.}

\item{dataset_name}{Name of the dataset (used as title).}
}
\value{
A girafe object.
}
\description{
Build an interactive population pyramid
}
