% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{CHAPTER_ROOTS}
\alias{CHAPTER_ROOTS}
\title{Anchor concept_ids for condition chapter assignment.}
\usage{
CHAPTER_ROOTS
}
\description{
Anchor concept_ids for condition chapter assignment.
}
\keyword{internal}
