% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_meta_by_sex}
\alias{compare_meta_by_sex}
\title{Meta-analysis across age groups: one row per concept_id x sex.}
\usage{
compare_meta_by_sex(meta_agegroups_df)
}
\arguments{
\item{meta_agegroups_df}{Output of \code{compare_meta_agegroups}.}
}
\value{
tibble with meta-analyzed rows.
}
\description{
Meta-analysis across age groups: one row per concept_id x sex.
}
