% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{CONDITION_RELATIONSHIPS}
\alias{CONDITION_RELATIONSHIPS}
\title{SNOMED relationship types to extract as condition attributes.}
\usage{
CONDITION_RELATIONSHIPS
}
\description{
SNOMED relationship types to extract as condition attributes.
}
\keyword{internal}
