% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\name{.build_caresite_sql}
\alias{.build_caresite_sql}
\title{Build the SELECT SQL for care-site cohort generation.}
\usage{
.build_caresite_sql(
  care_site_id,
  cohort_id,
  cdm_schema,
  collapse_strategy,
  restrict_to_observation
)
}
\description{
Build the SELECT SQL for care-site cohort generation.
}
\keyword{internal}
