% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{DRUG_RELATIONSHIPS}
\alias{DRUG_RELATIONSHIPS}
\title{Drug attribute relationships (empty - RxNorm relationships are structural).}
\usage{
DRUG_RELATIONSHIPS
}
\description{
Drug attribute relationships (empty - RxNorm relationships are structural).
}
\keyword{internal}
