% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_condition_attributes}
\alias{extract_condition_attributes}
\title{Extract condition SNOMED attributes.}
\usage{
extract_condition_attributes(cdm, condition_ids)
}
\arguments{
\item{cdm}{CDM reference object.}

\item{condition_ids}{Integer vector of concept_ids.}
}
\value{
tibble with columns: concept_id, relationship, target_concept_id, target_concept_name
}
\description{
Extract condition SNOMED attributes.
}
\keyword{internal}
