% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_condition_info}
\alias{extract_condition_info}
\title{Extract condition concept metadata.}
\usage{
extract_condition_info(cdm)
}
\arguments{
\item{cdm}{CDM reference object.}
}
\value{
tibble with columns: concept_id, concept_name, concept_code,
  vocabulary_id, concept_class_id, n_patients_total, n_records_total
}
\description{
Extract condition concept metadata.
}
\keyword{internal}
