% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_demographics}
\alias{extract_demographics}
\title{Extract demographics (birth year x sex).}
\usage{
extract_demographics(cdm)
}
\arguments{
\item{cdm}{CDM reference object.}
}
\value{
tibble with columns: sex, birth_year, patient_count
}
\description{
Extract demographics (birth year x sex).
}
\keyword{internal}
