% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_denominators}
\alias{extract_denominators}
\title{Extract the ACHILLES-116 denominator: persons observed per year x sex x age group.}
\usage{
extract_denominators(cdm)
}
\arguments{
\item{cdm}{CDM reference object.}
}
\value{
tibble with columns: year, sex, age_group, denominator
}
\description{
Extract the ACHILLES-116 denominator: persons observed per year x sex x age group.
}
\keyword{internal}
