% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_drug_chapters}
\alias{extract_drug_chapters}
\title{Extract drug chapter assignments (ATC 1st level).}
\usage{
extract_drug_chapters(cdm, drug_ids)
}
\arguments{
\item{cdm}{CDM reference object.}

\item{drug_ids}{Integer vector of Ingredient concept_ids.}
}
\value{
tibble with same columns as extract_condition_chapters.
}
\description{
Extract drug chapter assignments (ATC 1st level).
}
\keyword{internal}
