% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_drug_info}
\alias{extract_drug_info}
\title{Extract drug (ingredient) concept metadata.}
\usage{
extract_drug_info(cdm)
}
\arguments{
\item{cdm}{CDM reference object.}
}
\value{
tibble with same columns as extract_condition_info.
}
\description{
Extract drug (ingredient) concept metadata.
}
\keyword{internal}
