% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_drug_prevalence}
\alias{extract_drug_prevalence}
\title{Extract drug prevalence at the Ingredient level.}
\usage{
extract_drug_prevalence(cdm, denom_df)
}
\arguments{
\item{cdm}{CDM reference object.}

\item{denom_df}{Pre-computed denominator tibble.}
}
\value{
tibble with same columns as extract_condition_prevalence.
}
\description{
Extract drug prevalence at the Ingredient level.
}
\keyword{internal}
