% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_procedure_attributes}
\alias{extract_procedure_attributes}
\title{Extract procedure SNOMED attributes.}
\usage{
extract_procedure_attributes(cdm, procedure_ids)
}
\arguments{
\item{cdm}{CDM reference object.}

\item{procedure_ids}{Integer vector of concept_ids.}
}
\value{
tibble with same columns as extract_condition_attributes.
}
\description{
Extract procedure SNOMED attributes.
}
\keyword{internal}
