% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_procedure_chapters}
\alias{extract_procedure_chapters}
\title{Extract procedure chapter and sub-chapter assignments.}
\usage{
extract_procedure_chapters(cdm, procedure_ids)
}
\arguments{
\item{cdm}{CDM reference object.}

\item{procedure_ids}{Integer vector of concept_ids.}
}
\value{
tibble with same columns as extract_condition_chapters.
}
\description{
Extract procedure chapter and sub-chapter assignments.
}
\keyword{internal}
