% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_procedure_info}
\alias{extract_procedure_info}
\title{Extract procedure concept metadata.}
\usage{
extract_procedure_info(cdm)
}
\arguments{
\item{cdm}{CDM reference object.}
}
\value{
tibble with same columns as extract_condition_info.
}
\description{
Extract procedure concept metadata.
}
\keyword{internal}
