% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_procedure_prevalence}
\alias{extract_procedure_prevalence}
\title{Extract procedure prevalence.}
\usage{
extract_procedure_prevalence(cdm, denom_df)
}
\arguments{
\item{cdm}{CDM reference object.}

\item{denom_df}{Pre-computed denominator tibble.}
}
\value{
tibble with columns: concept_id, year, sex, age_group,
  patient_count, record_count, denominator, prevalence
}
\description{
Extract procedure prevalence.
}
\keyword{internal}
