% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{GENDER_LABELS}
\alias{GENDER_LABELS}
\title{Gender concept_id to label mapping.}
\format{
An object of class \code{character} of length 2.
}
\usage{
GENDER_LABELS
}
\description{
Gender concept_id to label mapping.
}
\keyword{internal}
