% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{K_ANONYMITY}
\alias{K_ANONYMITY}
\title{k-anonymity threshold. Any stratum cell with fewer patients is suppressed.}
\usage{
K_ANONYMITY
}
\description{
k-anonymity threshold. Any stratum cell with fewer patients is suppressed.
}
\keyword{internal}
