% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{list_comparisons}
\alias{list_comparisons}
\title{List available comparisons.}
\usage{
list_comparisons()
}
\value{
Character vector of comparison directory names.
}
\description{
List available comparisons.
}
