% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{load_comparison}
\alias{load_comparison}
\title{Load a previously saved comparison from CSV.}
\usage{
load_comparison(d1_name, d2_name)
}
\arguments{
\item{d1_name}{Dataset 1 name.}

\item{d2_name}{Dataset 2 name.}
}
\value{
Named list of tibbles.
}
\description{
Load a previously saved comparison from CSV.
}
