% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{PROCEDURE_CHAPTER_ROOTS}
\alias{PROCEDURE_CHAPTER_ROOTS}
\title{Anchor concept_ids for procedure chapter assignment.}
\usage{
PROCEDURE_CHAPTER_ROOTS
}
\description{
Anchor concept_ids for procedure chapter assignment.
}
\keyword{internal}
