% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{PROCEDURE_RELATIONSHIPS}
\alias{PROCEDURE_RELATIONSHIPS}
\title{SNOMED relationship types to extract as procedure attributes.}
\usage{
PROCEDURE_RELATIONSHIPS
}
\description{
SNOMED relationship types to extract as procedure attributes.
}
\keyword{internal}
