% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SOURCES_DIR}
\alias{SOURCES_DIR}
\title{Base directory for extracted source data.}
\format{
An object of class \code{character} of length 1.
}
\usage{
SOURCES_DIR
}
\description{
Base directory for extracted source data.
}
\keyword{internal}
