% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syrona-package.R
\docType{package}
\name{syrona-package}
\alias{syrona}
\alias{syrona-package}
\title{syrona: Stratified prevalence comparison across OMOP CDM datasets}
\description{
Derives stratified prevalence tables from the condition, procedure, and drug
records in OMOP CDM databases, computes log2 prevalence ratios between
paired datasets, and synthesizes them via random-effects meta-analysis
at multiple aggregation levels (year, age group, sex).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/HealthInformaticsUT/Syrona}
  \item Report bugs at \url{https://github.com/HealthInformaticsUT/Syrona/issues}
}

}
\author{
\strong{Maintainer}: Maarja Pajusalu \email{maarja.pajusalu@ut.ee}

Authors:
\itemize{
  \item Maarja Pajusalu \email{maarja.pajusalu@ut.ee}
}

}
\keyword{internal}
