% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{syrona_disconnect}
\alias{syrona_disconnect}
\title{Disconnect from an OMOP CDM database.}
\usage{
syrona_disconnect(db)
}
\arguments{
\item{db}{Connection list returned by \code{syrona_connect} or
\code{syrona_connect_pg}.}
}
\value{
No return value, called for its side effect of closing the database
  connection.
}
\description{
Disconnect from an OMOP CDM database.
}
