% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{truncate_name}
\alias{truncate_name}
\title{Truncate a concept name to a maximum number of characters}
\usage{
truncate_name(name, max = CONCEPT_NAME_MAX)
}
\arguments{
\item{name}{Character vector of names.}

\item{max}{Maximum character length (default CONCEPT_NAME_MAX).}
}
\value{
Character vector of (possibly truncated) names.
}
\description{
Adds "..." if truncated.
}
