% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{AGE_CLAMP_MAX}
\alias{AGE_CLAMP_MAX}
\title{Age decade clamping: all ages >= this value merge into a single "80+" group.}
\format{
An object of class \code{integer} of length 1.
}
\usage{
AGE_CLAMP_MAX
}
\description{
Age decade clamping: all ages >= this value merge into a single "80+" group.
}
\keyword{internal}
