% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{ATC_CHAPTER_VOCAB}
\alias{ATC_CHAPTER_VOCAB}
\title{ATC vocabulary constants for chapter lookup.}
\format{
An object of class \code{character} of length 1.
}
\usage{
ATC_CHAPTER_VOCAB
}
\description{
ATC vocabulary constants for chapter lookup.
}
\keyword{internal}
