% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CHAPTER_ROOTS}
\alias{CHAPTER_ROOTS}
\title{Anchor concept_ids for condition chapter assignment.}
\format{
An object of class \code{list} of length 2.
}
\usage{
CHAPTER_ROOTS
}
\description{
Anchor concept_ids for condition chapter assignment.
}
\keyword{internal}
