% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CONDITION_RELATIONSHIPS}
\alias{CONDITION_RELATIONSHIPS}
\title{SNOMED relationship types to extract as condition attributes.}
\format{
An object of class \code{character} of length 6.
}
\usage{
CONDITION_RELATIONSHIPS
}
\description{
SNOMED relationship types to extract as condition attributes.
}
\keyword{internal}
