% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{K_ANONYMITY}
\alias{K_ANONYMITY}
\title{k-anonymity threshold. Any stratum cell with fewer patients is suppressed.}
\format{
An object of class \code{integer} of length 1.
}
\usage{
K_ANONYMITY
}
\description{
k-anonymity threshold. Any stratum cell with fewer patients is suppressed.
}
\keyword{internal}
