% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{load_dataset}
\alias{load_dataset}
\title{Load a previously saved dataset from CSV.}
\usage{
load_dataset(dataset_name)
}
\arguments{
\item{dataset_name}{Short label (subfolder name under data/sources/).}
}
\value{
Named list of tibbles.
}
\description{
Load a previously saved dataset from CSV.
}
