% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{PROCEDURE_CHAPTER_ROOTS}
\alias{PROCEDURE_CHAPTER_ROOTS}
\title{Anchor concept_ids for procedure chapter assignment.}
\format{
An object of class \code{list} of length 2.
}
\usage{
PROCEDURE_CHAPTER_ROOTS
}
\description{
Anchor concept_ids for procedure chapter assignment.
}
\keyword{internal}
