% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{PROCEDURE_RELATIONSHIPS}
\alias{PROCEDURE_RELATIONSHIPS}
\title{SNOMED relationship types to extract as procedure attributes.}
\format{
An object of class \code{character} of length 2.
}
\usage{
PROCEDURE_RELATIONSHIPS
}
\description{
SNOMED relationship types to extract as procedure attributes.
}
\keyword{internal}
