% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{save_comparison}
\alias{save_comparison}
\title{Save comparison tables to CSV.}
\usage{
save_comparison(tables, d1_name, d2_name)
}
\arguments{
\item{tables}{Named list of tibbles.}

\item{d1_name}{Dataset 1 name.}

\item{d2_name}{Dataset 2 name.}
}
\description{
Save comparison tables to CSV.
}
\keyword{internal}
