% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{save_dataset}
\alias{save_dataset}
\title{Save extracted tables to CSV.}
\usage{
save_dataset(tables, dataset_name, db_path = NA_character_)
}
\arguments{
\item{tables}{Named list of tibbles.}

\item{dataset_name}{Short label for the dataset.}

\item{db_path}{Database path (stored in metadata).}
}
\description{
Save extracted tables to CSV.
}
\keyword{internal}
